induvidual_parts/Prompt_Optimizer/requirements.txt
openai
tiktoken
sentence-transformers
python-dotenv